requirements.txt
#nose
#zope.component[zcml]
nose
zope.schema
zope.i18nmessageid
sqlalchemy
